Move `LocalFeedLoader` and `FeedStore` collaborator to its own prod file
